renamed ProductSearch* classes to new namespace and removed prefix in class names
