+++ packages/framework/src/Model/Product/Elasticsearch/ExportScheduler.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrameworkBundle\Model\Product\Elasticsearch;
+
+use Shopsys\FrameworkBundle\Component\Elasticsearch\AbstractExportScheduler;
+
+class ExportScheduler extends AbstractExportScheduler
+{
+}
