--- packages/framework/src/Model/Product/Search/Export/ProductSearchExportScheduler.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\FrameworkBundle\Model\Product\Search\Export;
-
-use Shopsys\FrameworkBundle\Component\Elasticsearch\AbstractExportScheduler;
-
-class ProductSearchExportScheduler extends AbstractExportScheduler
-{
-}
