+++ packages/framework/src/Resources/config/services_test.yml
@@ -220,6 +220,6 @@ services:
 
     Shopsys\FrameworkBundle\Model\Product\Pricing\ProductInputPriceRecalculator: ~
 
-    Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportSubscriber:
+    Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ExportSubscriber:
         autoconfigure: false
