+++ packages/framework/tests/Unit/Model/Elasticsearch/IndexRegistryTest.php
@@ -8,8 +8,8 @@ use PHPUnit\Framework\TestCase;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Elasticsearch\Exception\ElasticsearchIndexException;
 use Shopsys\FrameworkBundle\Component\Elasticsearch\IndexRegistry;
+use Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ExportWithFilterRepository;
 use Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ProductIndex;
-use Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportWithFilterRepository;
 use Tests\FrameworkBundle\Unit\Model\Elasticsearch\__fixtures\CategoryIndex;
 
 class IndexRegistryTest extends TestCase
@@ -25,11 +25,11 @@ class IndexRegistryTest extends TestCase
 
         /** @var \Shopsys\FrameworkBundle\Component\Domain\Domain $domainMock */
         $domainMock = $this->createMock(Domain::class);
-        /** @var \Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportWithFilterRepository $productSearchExportWithFilterRepositoryMock */
-        $productSearchExportWithFilterRepositoryMock = $this->createMock(ProductSearchExportWithFilterRepository::class);
+        /** @var \Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ExportWithFilterRepository $exportWithFilterRepositoryMock */
+        $exportWithFilterRepositoryMock = $this->createMock(ExportWithFilterRepository::class);
 
         $this->indexRegistry = new IndexRegistry([
-            new ProductIndex($domainMock, $productSearchExportWithFilterRepositoryMock),
+            new ProductIndex($domainMock, $exportWithFilterRepositoryMock),
             new CategoryIndex(),
         ]);
     }
