+++ project-base/config/services_test.yml
@@ -58,7 +58,7 @@ services:
 
     Shopsys\FrameworkBundle\Model\Product\Availability\AvailabilityDataFactory: ~
 
-    Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportWithFilterRepository: ~
+    Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ExportWithFilterRepository: ~
 
     Shopsys\FrameworkBundle\Model\Product\Brand\BrandDataFactoryInterface: '@App\Model\Product\Brand\BrandDataFactory'
